Extract ignore methods into IgnoreMask
