+++ src/sqlfluff/core/linter/__init__.py
@@ -3,9 +3,9 @@
 from sqlfluff.core.linter.common import (
     RuleTuple,
     ParsedString,
-    NoQaDirective,
     RenderedFile,
 )
+from sqlfluff.core.linter.noqa import NoQaDirective, IgnoreMask
 from sqlfluff.core.linter.linted_file import LintedFile
 from sqlfluff.core.linter.linting_result import LintingResult
 from sqlfluff.core.linter.linter import Linter
@@ -14,6 +14,7 @@ __all__ = (
     "RuleTuple",
     "ParsedString",
     "NoQaDirective",
+    "IgnoreMask",
     "LintedFile",
     "LintingResult",
     "Linter",
